Modify shift
