+++ core/js/src/main/scala/cats/effect/internals/IOTimer.scala
@@ -17,6 +17,7 @@
 package cats.effect
 package internals
 
+import cats.effect.internals.IOShift.Tick
 import scala.concurrent.ExecutionContext
 import scala.concurrent.duration.{FiniteDuration, MILLISECONDS, NANOSECONDS, TimeUnit}
 import scala.scalajs.js
@@ -31,7 +32,7 @@ import scala.scalajs.js
  * introduces latency even when the specified delay is zero).
  */
 private[internals] class IOTimer extends Timer[IO] {
-  import IOTimer.{Tick, setTimeout, clearTimeout, setImmediateRef}
+  import IOTimer.{ScheduledTick, setTimeout, clearTimeout, setImmediateRef}
 
   final def clockRealTime(unit: TimeUnit): IO[Long] =
     IO(unit.convert(System.currentTimeMillis(), MILLISECONDS))
@@ -42,7 +43,9 @@ private[internals] class IOTimer extends Timer[IO] {
   final def sleep(timespan: FiniteDuration): IO[Unit] =
     IO.Async(new IOForkedStart[Unit] {
       def apply(conn: IOConnection, cb: Either[Throwable, Unit] => Unit): Unit = {
-        val task = setTimeout(timespan.toMillis, new Tick(conn, cb))
+        val task = setTimeout(timespan.toMillis, new ScheduledTick(conn, cb))
+        // On the JVM this would need a ForwardCancelable,
+        // but not on top of JS as we don't have concurrency
         conn.push(() => clearTimeout(task))
       }
     })
@@ -50,7 +53,7 @@ private[internals] class IOTimer extends Timer[IO] {
   final def shift: IO[Unit] =
     IO.Async(new IOForkedStart[Unit] {
       def apply(conn: IOConnection, cb: Callback.T[Unit]): Unit = {
-        execute(new Tick(null, cb))
+        execute(new Tick(cb))
       }
     })
 
@@ -100,13 +103,13 @@ private[internals] object IOTimer {
       js.Dynamic.global.setTimeout
   }
 
-  private final class Tick(
+  private final class ScheduledTick(
     conn: IOConnection,
     cb: Either[Throwable, Unit] => Unit)
     extends Runnable {
 
     def run() = {
-      if (conn ne null) conn.pop()
+      conn.pop()
       cb(Callback.rightUnit)
     }
   }
